// src/models/giocata.ts
export class Giocata {

    event_id: number;
    hometeam_name: string;
    awayteam_name: string;
    margin_live: number;
    quota_analizzata: number;
    bookmaker: string;
    x_id: number;
    segno_analizzato: string;
    event_time: string;
      numeroGiocateTrovate: number;
      previsioneMarginMax: number; 
      previsioneMarginMaxRelativo: number; 
      previsioneMarginEnd: number;
      previsioneQuotaLast: number;

  }